Finalisation Projet

+++ backend/controllers/sauce.js
@@ -34,18 +34,24 @@ exports.updateSauce = (req, res, next) => {
   const sauceObject = req.file ? {
     ...JSON.parse(req.body.sauce),
   imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`} : { ...req.body }
-  delete sauceObject._userId;
           Sauce.findOne({_id: req.params.id})
             .then((sauce) => {
               if (sauce.userId != req.auth.userId) {
-                  res.status(401).json({ message : 'Not authorized'});
+                  res.status(403).json({ message : 'Not authorized'});
               } else {
-                const filename = sauce.imageUrl.split("/images/")[1]
-                fs.unlink(`images/${filename}`, () => {
+                if(req.file){
+                  const filename = sauce.imageUrl.split("/images/")[1]
+                    fs.unlink(`images/${filename}`, () => {
                     Sauce.updateOne({ _id : req.params.id}, {...sauceObject, _id: req.params.id})
                       .then(res.status(200).json({ message : "Sauce modifiée"}))
                       .catch(error => res.status(400).json({ error }))
                     })
+                    }
+                else{
+                  Sauce.updateOne({ _id : req.params.id}, {...sauceObject, _id: req.params.id})
+                      .then(res.status(200).json({ message : "Sauce modifiée"}))
+                      .catch(error => res.status(400).json({ error }))
+                }
                   }
                 })
 };
@@ -54,7 +60,7 @@ exports.deleteSauce = (req, res, next) => {
       Sauce.findOne({ _id : req.params.id })
           .then(sauce => {
               if (sauce.userId != req.auth.userId) {
-                  res.status(401).json({message: 'Not authorized'});
+                  res.status(403).json({message: 'Not authorized'});
               } else {
                 const filename = sauce.imageUrl.split("/images/")[1]
                   fs.unlink(`images/${filename}`, () => {
